[ESWE-1249] Resend employers; JB API client
- JB API client: call API to get all employers
