+++ src/integrationTest/kotlin/uk/gov/justice/digital/hmpps/jobsboardintegrationapi/integration/shared/infrastructure/JobsBoardApiWebClientShould.kt
@@ -5,11 +5,14 @@ import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
+import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.amazon
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.sainsburys
+import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.tesco
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.integration.IntegrationTestBase
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.integration.wiremock.HmppsAuthApiExtension.Companion.hmppsAuth
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.integration.wiremock.JobsBoardApiExtension.Companion.jobsBoardApi
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.shared.infrastructure.JobsBoardApiWebClient
+import kotlin.math.ceil
 
 class JobsBoardApiWebClientShould : IntegrationTestBase() {
 
@@ -41,6 +44,62 @@ class JobsBoardApiWebClientShould : IntegrationTestBase() {
       val actualEmployer = jobsBoardApiWebClient.getEmployer(employerId)
 
       assertThat(actualEmployer).isNull()
+    }
+  }
+
+  @Nested
+  @DisplayName("JobsBoard `GET` /employers")
+  inner class EmployersRetrieveAllEndpoint {
+    private val someEmployers =
+      arrayOf(sainsburys, tesco, amazon).map { it.copy(createdAt = timeProvider.nowAsInstant()) }.toTypedArray()
+
+    @Test
+    fun `return all employers`() {
+      hmppsAuth.stubGrantToken()
+      jobsBoardApi.stubRetrieveAllEmployers(*someEmployers)
+
+      val employersResponse = jobsBoardApiWebClient.getAllEmployers()
+
+      assertThat(employersResponse.content.size).isEqualTo(someEmployers.size)
+      assertThat(employersResponse.content).containsOnly(*someEmployers)
+    }
+
+    @Test
+    fun `return empty list, when no employer exist`() {
+      hmppsAuth.stubGrantToken()
+      jobsBoardApi.stubRetrieveAllEmployers()
+
+      val employersResponse = jobsBoardApiWebClient.getAllEmployers()
+
+      assertThat(employersResponse.content).isEmpty()
+      employersResponse.page.run {
+        assertThat(number).isEqualTo(0)
+        assertThat(totalElements).isEqualTo(0)
+        assertThat(totalPages).isEqualTo(0)
+      }
+    }
+
+    @Test
+    fun `return a page of employers`() {
+      hmppsAuth.stubGrantToken()
+      val currentPage = 1
+      val pageSize = 1
+      val expectedTotalElements = someEmployers.size.toLong()
+      val expectedEmployer = someEmployers[1]
+      val expectedTotalPages = ceil(expectedTotalElements.toDouble() / pageSize).toInt()
+      jobsBoardApi.stubRetrieveAllEmployers(currentPage, pageSize, expectedTotalElements, expectedEmployer)
+
+      val employersResponse = jobsBoardApiWebClient.getAllEmployers(currentPage, pageSize)
+
+      assertThat(employersResponse.content)
+        .hasSize(1)
+        .first().isEqualTo(expectedEmployer)
+      employersResponse.page.run {
+        assertThat(number).isEqualTo(currentPage)
+        assertThat(size).isEqualTo(pageSize)
+        assertThat(totalPages).isEqualTo(expectedTotalPages)
+        assertThat(totalElements).isEqualTo(expectedTotalElements)
+      }
     }
   }
 }
