+++ src/test/kotlin/uk/gov/justice/digital/hmpps/jobsboardintegrationapi/shared/infrastructure/JobsBoardApiWebClientShould.kt
@@ -5,8 +5,11 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.junit.jupiter.MockitoExtension
+import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.amazon
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.sainsburys
+import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.employers.domain.EmployerObjects.tesco
 import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.jobs.domain.JobObjects.tescoWarehouseHandler
+import uk.gov.justice.digital.hmpps.jobsboardintegrationapi.shared.infrastructure.JobsBoardApiWebClient.Companion.employerResponseTypeRef
 
 @ExtendWith(MockitoExtension::class)
 class JobsBoardApiWebClientShould : JobsBoardApiWebClientTestBase() {
@@ -14,7 +17,9 @@ class JobsBoardApiWebClientShould : JobsBoardApiWebClientTestBase() {
   @Nested
   inner class GivenAnEmployer {
     private val employer = sainsburys.copy(createdAt = defaultCurrentTime)
-    private val uri = "/employers/{id}"
+    private val someEmployers = arrayOf(sainsburys, tesco, amazon)
+    private val baseUri = "/employers"
+    private val uri = "$baseUri/{id}"
 
     @Test
     fun `return employer details with a valid employer ID`() {
@@ -25,7 +30,19 @@ class JobsBoardApiWebClientShould : JobsBoardApiWebClientTestBase() {
       assertThat(actualEmployer).isEqualTo(employer)
     }
 
+    @Test
+    fun `return all employers`() {
+      replyOnGetEmployers(GetEmployersResponse.from(*someEmployers))
+      val employersPage = jobsBoardApiWebClient.getAllEmployers()
+
+      assertThat(employersPage.content)
+        .hasSize(someEmployers.size)
+        .contains(*someEmployers)
+    }
+
     private fun replyOnGetEmployerById(response: GetEmployerResponse, employerId: String) = replyOnRequestById(response.javaClass, response, uri, employerId)
+
+    private fun replyOnGetEmployers(response: GetEmployersResponse) = replyOnPagedRequest(employerResponseTypeRef, response, baseUri)
   }
 
   @Nested
